Change ``_get_rendered`` behavior for automatic generation of rendered
Duaring loaddata, ``pre_save`` method is not called that's why if you
want to create DEMO value or whatever with Fixture, empty field is
created.

To slove this issue, I change ``_get_rendered`` behavior to generate
rendered text automatically when stored rendered text is empty.

With this change, you can use fixture to create (but not update) field.