Tip for usage with default filter II.
Hi James!
As you state in the README, "rendered" attribute on Markup is of SafeText type. If "default" filter is applied to "rendered", output for browser is escaped - it displays HTML value, not interpret it as HTML. I've discovered that if I finish filter chain with "safe", HTML is not escaped.

Please keep in mind that I'm Django newbie and I'm unsure whether it's a bug or my misunderstanding of Django.